requirements.txt
streamlit
numpy
Pillow
matplotlib
h5py
opencv-python
scikit-learn